fix: streams import
