+++ my_caesar_cli.js
@@ -1,10 +1,8 @@
 import { pipeline } from 'stream';
 import handleError from './src/handle-error.js';
-import {
-  readStream,
-  writeStream,
-  transformStreams,
-} from './src/streams/streams.js';
+import createReadStream from './src/streams/read-stream.js';
+import createTransformStreams from './src/streams/transform-streams.js';
+import createWriteStream from './src/streams/write-stream.js';
 import validateOptions from './src/validate-options.js';
 
 try {
@@ -12,6 +10,10 @@ try {
 } catch (err) {
   handleError(err);
 }
+
+const readStream = createReadStream();
+const writeStream = createWriteStream();
+const transformStreams = createTransformStreams();
 
 pipeline(readStream, ...transformStreams, writeStream, (error) => {
   error ? console.log(console.error()) : console.log('finished');
